(Chore): Sync fork with upstream
As part of this, we have to change the channel names used to prevent a
conflict when the customer uses this package alongside the main
flutter_inappwebview package in the same app.

Alongside this, for the windows package, we also change the namespace and
macros. On Android we have to change the package name too.

Signed-off-by: Victor Otieno <victor@slick.co.ke>